composable::about

Module changelog

Changelog

Please keep one empty line before and after all headers. (This is required for git to produce a conflict when a release is made while a PR is open and the PR’s changelog entry would go into the wrong section).

And please only add new entries to the top of this list, right below the # Unreleased header.

The format is based on Keep a Changelog/Common Changelog, and this project adheres to Semantic Versioning.

Unreleased

Added

A binary target has been added to covert SVG files into Rust functions that can be used to render them as Shapes.

Dependency::get_or_default() has been added for Copy types.

Changed

Breaking: Dependency::new() is now Dependency::get().

Removed

Breaking: view has been moved into their own crate (in the workspace). It is still considered an unstable feature and is not considered with respect to the Semantic Versioning of this project.

Breaking: dependencies has been moved into their own crate (in the workspace).

Fixed

Added a default stack size to Store threads. The same thread stack size is now used on all platforms.

Store has been completely rebuilt over an explicit Reactor implementation. This enables it to support the needs of the Scheduler trait without spawning additional threads.

0.6.0 - 2024-07-22

Added

Asynchronous Effects that where removed in version 0.5 have been restored. They now run in a Local Async Executor, rather than a multi-threaded one,

The views feature is gated behind a new unstable feature flag. Development on Views may now continue without causing SemVer issues.
Unstable features are not considered when determining versioning.

Changed

Breaking: All traits and structs have been redesigned around the return_position_impl_trait_in_trait feature.

Breaking: Effects::send is now Effects::action; as actions, futures , and streams are the main types of Effects.

View drawing is governed by an Output trait; sending geometry to the GPU is now just one of the options available.

Gesture states are now expressed as .active, .hover, .focus to match the W3C’s user action pseudo classes as they are more widely known than the terms used in original Immediate-Mode Graphical User Interfaces presentation by Casey Muratori.

The Inter font has been updated to version 4.0.

Removed

Views are being rebuilt to better reflect the (new) separation between features:

views: the building blocks for creating UI elements around Actions, State and Reducers.

default_ui: a useful set of pre-built UI elements usable as-is, or as examples of how to build more complex elements.

Fixed

Better documentation and automated testing throughout.

Shape plans are cached for text layout.