Trait composable::derive_macros::RecursiveReducer

pub trait RecursiveReducer {
    type Action;

    // Required method
    fn reduce(&mut self, action: Self::Action, send: impl Effects<Self::Action>);
}

See the RecursiveReducer macro for example usage.

Required Associated Types

type Action

All of the possible actions that can be used to modify state. Equivalent to Reducer::Action.

Required Methods

fn reduce(&mut self, action: Self::Action, send: impl Effects<Self::Action>)

This reduce should perform any actions that are needed before the macro recurses into the other reducers.

Object Safety

This trait is not object safe.