Struct composable::TestStore

pub struct TestStore<State: Reducer>
where
    <State as Reducer>::Action: Debug,
{ /* private fields */ }

A state container for the application testing.

Example

Here is the second Reducer example being tested with a TestStore.

#[derive(Clone, Debug, Default, PartialEq)]
struct State {
    n: usize,
}

#[derive(Debug, PartialEq)]
enum Action {
    Increment,
    Decrement,
}

use Action::*;
impl Reducer for State {
    type Action = Action;
    type Output = Self;

    // This reducer ensures the value is always an even number
    fn reduce(&mut self, action: Action, send: impl Effects<Action>) {
        match action {
            Increment => {
                self.n += 1;
                if self.n % 2 == 1 {
                    send.action(Increment);
                }
            }
            Decrement => {
                self.n -= 1;
                if self.n % 2 == 1 {
                    send.action(Decrement);
                }
            }
        }
    }
}

let mut store = TestStore::<State>::default();

store.send(Increment, |state| state.n = 1);
store.recv(Increment, |state| state.n = 2);

store.send(Increment, |state| state.n = 3);
store.recv(Increment, |state| state.n = 4);

store.send(Decrement, |state| state.n = 3);
store.recv(Decrement, |state| state.n = 2);

let n = store.into_inner().n;
assert_eq!(n, 2);

Implementations

impl<State: Reducer> TestStore<State>
where <State as Reducer>::Action: Debug,

pub fn new<F>(with: F) -> Self
where F: FnOnce() -> State,

Creates a new Store with its initial state generated by a function.

pub fn with_initial(state: State) -> Self

Creates a new Store with state as its initial state.

pub fn send( &mut self, action: <State as Reducer>::Action, assert: impl FnOnce(&mut State), )
where State: Clone + Debug + PartialEq, <State as Reducer>::Action: 'static,

Calls the Store’s Reducer with action and asserts the expected state changes.

pub fn recv( &mut self, action: <State as Reducer>::Action, assert: impl FnOnce(&mut State), )
where State: Clone + Debug + PartialEq, <State as Reducer>::Action: Debug + PartialEq + 'static,

Checks that the Store’s Reducer was called with action and asserts the expected state changes.

pub fn wait(&mut self)

Waits until all scheduled tasks have completed.

A timeout should be added to tests calling wait() to ensure that it does not wait forever if there are bugs in the asynchronous tasks. For example

ntest provides a timeout attribute macro, while

divan has a max_time parameter for its bench macro.

pub fn into_inner(self) -> <State as Reducer>::Output
where State: Into<<State as Reducer>::Output>,

Consumes the Store and returns its current state value.