composable

Trait Effects

pub trait Effects:
    Clone
    + Scheduler<Item = Self::Action>
    + 'static {
    type Action;

    // Required methods
    fn action(&self, action: impl Into<<Self as Effects>::Action>);
    fn stream<S: Stream<Item = <Self as Effects>::Action> + 'static>(
        &self,
        stream: S,
    );
    fn future<F: Future<Output = Option<<Self as Effects>::Action>> + 'static>(
        &self,
        future: F,
    )
       where <Self as Effects>::Action: 'static;
    fn task<S: Stream<Item = <Self as Effects>::Action> + 'static>(
        &self,
        stream: S,
    ) -> Task;

    // Provided method
    fn scope<ChildAction>(&self) -> Scoped<Self, ChildAction>
       where Self::Action: From<ChildAction> { ... }
}

Required Associated Types

type Action

Required Methods

fn action(&self, action: impl Into<<Self as Effects>::Action>)

An effect that sends an Action through the Store’s Reducer.

fn stream<S: Stream<Item = <Self as Effects>::Action> + 'static>( &self, stream: S, )

An effect that runs a Stream and sends every Action it returns through the Store’s Reducer.

fn future<F: Future<Output = Option<<Self as Effects>::Action>> + 'static>( &self, future: F, )
where <Self as Effects>::Action: 'static,

An effect that runs a Future and, if it returns an Action, sends it through the Store’s Reducer.

fn task<S: Stream<Item = <Self as Effects>::Action> + 'static>( &self, stream: S, ) -> Task

A Task represents asynchronous work that will then send zero or more Actions back into the Store’s Reducer as it runs.

Use this method if you need to ability to cancel the task while it is running. Otherwise future or stream should be preferred.

Provided Methods

fn scope<ChildAction>(&self) -> Scoped<Self, ChildAction>
where Self::Action: From<ChildAction>,

Scopes the Effects down to one that sends child actions.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".