composable

Trait Reducer

pub trait Reducer {
    type Action;

    // Required method
    fn reduce(
        &mut self,
        action: Self::Action,
        send: impl Effects<Action = Self::Action>,
    );
}

Reducers are responsible for updating a Store’s state in response to its Actions.

Required Associated Types

type Action

All of the possible actions that can be used to modify state.

Required Methods

fn reduce( &mut self, action: Self::Action, send: impl Effects<Action = Self::Action>, )

Updates the Reducer’s state in response to the action received.

Additional Actions that need to be performed as a side-effect of an Action should be invoked on send. The logic of the feature is performed by mutating its current State with Actions.

#[derive(Clone, Debug, Default, PartialEq)]
struct State {
    n: usize,
}

#[derive(Debug, PartialEq)]
enum Action {
    Increment,
    Decrement,
}

This is most easily done by implementing the Reducer trait directly on it’s State.

use Action::*;
impl Reducer for State {
    type Action = Action;

    fn reduce(&mut self, action: Action, _send: impl Effects<Action=Self::Action>) {
        match action {
            Increment => {
                self.n += 1;
            }
            Decrement => {
                self.n -= 1;
            }
        }
    }
}

The reduce method’s first responsibility is to mutate the feature’s current state given an action. Its second responsibility is to trigger effects that feed their actions back into the system. Currently reduce does not need to run any Effects so _send goes unused.

If the action does need side effects, then more would need to be done. For example, if reduce always maintained an even number for the State, then each Increment and Decrement would need an effect to follow:1

use Action::*;
impl Reducer for State {
    type Action = Action;

    // This reducer ensures the value is always an even number
    fn reduce(&mut self, action: Action, send: impl Effects<Action=Self::Action>) {
        match action {
            Increment => {
                self.n += 1;
                if self.n % 2 == 1 {
                    send.action(Increment); // ⬅︎
                }
            }
            Decrement => {
                self.n -= 1;
                if self.n % 2 == 1 {
                    send.action(Decrement); // ⬅︎
                }
            }
        }
    }
}

See TestStore for a more complete test of this example.

See Effects for all of the effects that can be used within a Reducer.

Granted, real code could just adjust the values by two. It is a contrived example to show how to use Effects, after all. ↩

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".