Module composable::views::ui::spacing

Available on crate features unstable and views and default_ui only.

Constants

L

The amount of space taken up by a spacer::L()

M

The amount of space taken up by a spacer::M()

S

The amount of space taken up by a spacer::S()

XL

The amount of space taken up by a spacer::XL()

XS

The amount of space taken up by a spacer::XS()

XXL

The amount of space taken up by a spacer::XXL()

XXS

The amount of space taken up by a spacer::XXS()

XXXL

The amount of space taken up by a spacer::XXXL()